Basic battle system working
I added the basic features for the player's first turn in the battle system and the ability to roll for initiative at the start as well as kill the enemy and celebrate it. But the system can't currently handle a situation where the first attack does not immediately kill the enemy. Also, the enemy's first turn is not yet implemented.